requirements.txt
pywinauto
PyYAML
